# api/openapi.yaml
openapi: 3.1.0
info:
  title: ApplyFlow API — Foundation and Application Studio
  version: 0.6.0
  description: Contract-first foundation and Studio data lifecycle. 40 operations are implemented, including personal AI settings and fixed-revision PDF/DOCX export;
    check x-implementation on each operation for availability. All mutations enforce CSRF + Origin; private responses
    use no-store. Default JSON limit 128 KiB; resume fact confirmation and document content writes 256
    KiB. Multipart file limit 10 MiB (proxy permits framing overhead). Text also has UTF-8 byte and semantic
    limits. Generation/refinement supports explicit mock execution only in this slice; real AI generation,
    billing, full profiles, interviews and analytics remain deferred. SSRF, file validation and lifecycle
    transactions are required before implementing routes.
servers:
- url: /
  description: Same-origin frontend and API
security:
- sessionCookie: []
...
paths:
  /api/auth/csrf:
    get:
      operationId: getCSRFToken
      x-implementation: implemented
      tags:
      - Authentication
      summary: Get signed CSRF context
      description: Available before login. Sets an anonymous CSRF context cookie, or binds to the current
        session. Never caches. This endpoint does not issue an authentication session.
      responses:
        '200':
          description: Success
          headers:
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            Set-Cookie:
              description: Signed CSRF context cookie, Secure on HTTPS; Path=/; SameSite=Lax.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CSRFToken'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/Unavailable'
      security: []
  /api/auth/register:
    post:
      operationId: register
      x-implementation: implemented
      tags:
      - Authentication
      summary: Create account and session
      description: Normalize email with trim+lowercase; validate full email in the handler. In one transaction
        create user and empty profile (version 1). Sets authentication cookie; no email verification/reset
        workflow in this slice.
      responses:
        '201':
          description: Success
          headers:
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            Set-Cookie:
              description: Sets/clears applyflow_session HttpOnly; Secure in HTTPS; SameSite=Lax; Path=/.
                Auth tokens never appear in JSON. Login/register rotate the CSRF context; fetch /api/auth/csrf
                again.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/Unavailable'
      security: []
      parameters:
      - name: X-CSRF-Token
        in: header
        required: true
        schema:
          type: string
          minLength: 16
        description: Signed CSRF token from GET /api/auth/csrf; also enforce same-origin Origin policy.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credentials'
      x-max-body-bytes: 131072
  /api/auth/login:
    post:
      operationId: login
      x-implementation: implemented
      tags:
      - Authentication
      summary: Log in
      description: Log in
      responses:
        '200':
          description: Success
          headers:
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            Set-Cookie:
              description: Sets/clears applyflow_session HttpOnly; Secure in HTTPS; SameSite=Lax; Path=/.
                Auth tokens never appear in JSON. Login/register rotate the CSRF context; fetch /api/auth/csrf
                again.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/Unavailable'
      security: []
      parameters:
      - name: X-CSRF-Token
        in: header
        required: true
        schema:
          type: string
          minLength: 16
        description: Signed CSRF token from GET /api/auth/csrf; also enforce same-origin Origin policy.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginCredentials'
      x-max-body-bytes: 131072
  /api/auth/logout:
    post:
      operationId: logout
      x-implementation: implemented
      tags:
      - Authentication
      summary: Clear session cookie
      description: Idempotently clears the cookie, including if the JWT expired. Validate signed CSRF
        context and Origin. Does not revoke a copied stateless JWT; no refresh tokens in this slice.
      responses:
        '204':
          description: Success
          headers:
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            Set-Cookie:
              description: Sets/clears applyflow_session HttpOnly; Secure in HTTPS; SameSite=Lax; Path=/.
                Auth tokens never appear in JSON. Login/register rotate the CSRF context; fetch /api/auth/csrf
                again.
              schema:
                type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/Unavailable'
      security: []
      parameters:
      - name: X-CSRF-Token
        in: header
        required: true
        schema:
          type: string
          minLength: 16
        description: Signed CSRF token from GET /api/auth/csrf; also enforce same-origin Origin policy.
  /api/auth/me:
    get:
      operationId: getCurrentSession
      x-implementation: implemented
      tags:
      - Authentication
      summary: Get current identity and expiry
      description: Get current identity and expiry
      responses:
        '200':
          description: Success
          headers:
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/Unavailable'
...
  /api/files:
    post:
      operationId: uploadPrivateFile
      x-implementation: implemented
      tags:
      - Files
      summary: Upload one private source file
      description: 'One multipart PDF/DOCX base_resume file <=10 MiB; total body <=11 MiB. Validate actual
        content. PDF must be readable, unencrypted and <=20 pages. PDF rejection codes include invalid_pdf,
        pdf_restricted and pdf_page_limit (415); validator execution failure uses file_validator_unavailable (503). DOCX: <=512 ZIP entries, <=8 MiB per
        entry and <=32 MiB decompressed total; no macros, embeddings or external non-hyperlink relationships.
        DOCX pagination is renderer-dependent and not checked at upload. Screenshot upload is deferred.
        Ready denotes validated private storage, not parsed facts. No storage key returned. Unknown upload
        result is not automatically retried; orphan cleanup is not yet implemented.'
      parameters:
      - name: X-CSRF-Token
        in: header
        required: true
        schema:
          type: string
          minLength: 16
        description: Signed CSRF token from GET /api/auth/csrf; also enforce same-origin Origin policy.
      responses:
        '201':
          description: Success; persisted before responding.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            Location:
              description: Same-origin URL of accepted resource or task.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrivateFile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/Unavailable'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/FileUpload'
      x-max-body-bytes: 11534336
  /api/files/{id}/download:
    get:
      operationId: downloadPrivateFile
      x-implementation: implemented
      tags:
      - Files
      summary: Download an owned ready file
      description: Download an owned ready file
      parameters:
      - name: id
        in: path
        required: true
        schema: &id001
          type: string
          format: uuid
      responses:
        '200':
          description: Success; persisted before responding.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            Content-Disposition:
              schema:
                type: string
              description: attachment; filename sanitized, UTF-8 filename* supported
            X-Content-Type-Options:
              schema:
                type: string
                const: nosniff
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/Unavailable'
  /api/workspaces:
    get:
      operationId: listWorkspaces
      x-implementation: implemented
      tags:
      - Workspaces
      summary: List active workspaces
      description: Cursor pagination by updated_at,id descending. Owner from session; exclude archived.
        Cursor bound to filters; no total guarantee.
      parameters:
      - name: limit
        in: query
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
      - name: cursor
        in: query
        schema:
          type: string
          minLength: 1
          maxLength: 2048
      responses:
        '200':
          description: Success; persisted before responding.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/Unavailable'
    post:
      operationId: createWorkspace
      x-implementation: implemented
      tags:
      - Workspaces
      summary: Create a recoverable workspace
      description: Create a recoverable workspace
      parameters:
      - name: X-CSRF-Token
        in: header
        required: true
        schema:
          type: string
          minLength: 16
        description: Signed CSRF token from GET /api/auth/csrf; also enforce same-origin Origin policy.
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          format: uuid
        description: One UUID per user operation. Retry identical body with same key. Replay saved response
          before version checks; different body returns 409. Stored at least 24h and while tasks are nonterminal.
      responses:
        '201':
          description: Success; persisted before responding.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            Location:
              description: Same-origin URL of accepted resource or task.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/Unavailable'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkspaceCreate'
      x-max-body-bytes: 131072
  /api/workspaces/{id}:
    get:
      operationId: getWorkspace
      x-implementation: implemented
      tags:
      - Workspaces
      summary: Read workspace including archived
      description: Read workspace including archived
      parameters:
      - name: id
        in: path
        required: true
        schema: *id001
      responses:
        '200':
          description: Success; persisted before responding.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/Unavailable'
    patch:
      operationId: updateWorkspace
      x-implementation: implemented
      tags:
      - Workspaces
      summary: Update workspace with optimistic version
      description: Owner/version CAS. Ready confirmed owned resume required. Archived workspace returns
        409. Changing selection never rewrites historical runs.
      parameters:
      - name: id
        in: path
        required: true
        schema: *id001
      - name: X-CSRF-Token
        in: header
        required: true
        schema:
          type: string
          minLength: 16
        description: Signed CSRF token from GET /api/auth/csrf; also enforce same-origin Origin policy.
      responses:
        '200':
          description: Success; persisted before responding.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/Unavailable'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkspacePatch'
      x-max-body-bytes: 131072
  /api/workspaces/{id}/sources:
    post:
      operationId: createJobSource
      x-implementation: implemented
      tags:
      - Workspaces
      summary: Submit text, URL or screenshot JD
      description: Text-only mock slice. Atomically persist immutable text, extraction task/outbox and
        replay response; update workspace version/current_source_id and clear job_revision_id. The Worker
        returns the original text with null company/role for manual confirmation; no AI extraction is
        claimed. Images/URLs are not accepted yet. Text is limited to 65536 UTF-8 bytes.
      parameters:
      - name: id
        in: path
        required: true
        schema: *id001
      - name: X-CSRF-Token
        in: header
        required: true
        schema:
          type: string
          minLength: 16
        description: Signed CSRF token from GET /api/auth/csrf; also enforce same-origin Origin policy.
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          format: uuid
        description: One UUID per user operation. Retry identical body with same key. Replay saved response
          before version checks; different body returns 409. Stored at least 24h and while tasks are nonterminal.
      responses:
        '202':
          description: Success; persisted before responding.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            Location:
              description: Same-origin URL of accepted resource or task.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceAccepted'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/Unavailable'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SourceCreate'
      x-max-body-bytes: 131072
  /api/workspaces/{id}/sources/{source_id}:
    get:
      operationId: getJobSource
      x-implementation: implemented
      tags:
      - Workspaces
      summary: Read submitted source
      description: Read submitted source
      parameters:
      - name: id
        in: path
        required: true
        schema: *id001
      - name: source_id
        in: path
        required: true
        schema: *id001
      responses:
        '200':
          description: Success; persisted before responding.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobSource'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/Unavailable'
  /api/workspaces/{id}/job-revisions:
    post:
      operationId: confirmJobSource
      x-implementation: implemented
      tags:
      - Workspaces
      summary: Confirm extracted job details
      description: Owner/version check; source must be current. Reject blank fields and JD over 65536
        UTF-8 bytes. Insert immutable revision and update workspace selection/version in one transaction.
      parameters:
      - name: id
        in: path
        required: true
        schema: *id001
      - name: X-CSRF-Token
        in: header
        required: true
        schema:
          type: string
          minLength: 16
        description: Signed CSRF token from GET /api/auth/csrf; also enforce same-origin Origin policy.
      responses:
        '201':
          description: Success; persisted before responding.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            Location:
              description: Same-origin URL of accepted resource or task.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobConfirmed'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/Unavailable'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobConfirm'
      x-max-body-bytes: 131072
  /api/workspaces/{id}/job-revisions/{revision_id}:
    get:
      operationId: getJobRevision
      x-implementation: implemented
      tags:
      - Workspaces
      summary: Read confirmed job revision
      description: Read confirmed job revision
      parameters:
      - name: id
        in: path
        required: true
        schema: *id001
      - name: revision_id
        in: path
        required: true
        schema: *id001
      responses:
        '200':
          description: Success; persisted before responding.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobRevision'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/Unavailable'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: applyflow_session
      description: Short-lived signed JWT; HttpOnly, Secure on HTTPS, SameSite=Lax, Path=/.
  schemas:
...
    CSRFToken:
      type: object
      additionalProperties: false
      properties:
        csrf_token:
          type: string
          minLength: 16
        expires_at:
          type: string
          format: date-time
      required:
      - csrf_token
      - expires_at
...
  responses:
    BadRequest:
      description: Malformed JSON, unknown fields, invalid values or query/cursor.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/X-Request-ID'
        Cache-Control:
          $ref: '#/components/headers/Cache-Control'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  headers:
    X-Request-ID:
      description: Server-generated request correlation ID.
      schema:
        type: string
    Cache-Control:
      description: Private responses always use no-store.
      schema:
        type: string
        const: no-store